PL Scenarios 19+20 - Work on #37 - Work on #27 - Recompile maps with new elevation transition flag

+++ View/Edit/EditPane.cs
@@ -179,7 +179,7 @@ namespace PanzerBlitz
 			}
 			else if (_ModeSelect.Value.Value == _ElevationPage)
 			{
-				Tile.Configuration.SetElevation(Tile.Configuration.Elevation + 1);
+				Tile.Configuration.SetElevation((byte)(Tile.Configuration.Elevation + 1));
 			}
 			else if (_ModeSelect.Value.Value == _MapRegionPage)
 			{
@@ -213,7 +213,7 @@ namespace PanzerBlitz
 			}
 			else if (_ModeSelect.Value.Value == _ElevationPage)
 			{
-				Tile.Configuration.SetElevation(Tile.Configuration.Elevation - 1);
+				Tile.Configuration.SetElevation((byte)(Tile.Configuration.Elevation - 1));
 			}
 			else if (_ModeSelect.Value.Value == _MapRegionPage)
 			{
